Stretched Images Fix
This one took me a little while, but I found a solution that works great. Currently, it's using a sclae factor of 1.5 for smaller images to be enlarged. You can change this to a bigger number if you want the smaller images to be enlargered to a greater size.
Fixes #28