Update 10.20
I refactored and worked with the loading of scripts some more which has lead to the best results for the userscript. There is also some new code added that I am planning to use for an upcoming feature. I just have to finalize some of the data handling and selection of the storage. I'm surprise no one has done this before : ).