A working version of forum saver
This is a working version of the forum post text saver. It is currently buggy and requires work to reduce the issues caused by the different scenerios that can happen. I will later add auto saving and other features. This still needs a lot of work, but it should be a very nice system once complete.