rebuilt custom scroll with new element
I added new dom-creation which is now an offical package. I changed everything to comply with its standards and made sure to fix the bugs with it. It also helped me condense the code for building the scrollers on the front page. I do plan to use this and more basic operations instead of jQuery due to performace concerns