Conversion and scrolls
Added the last two needed scrolls to the front page. I also started converting the jquery to vanilla JS due to performance and bloat. I want to scale back jquery as much as possible due to it being a very heavy when compared to modern implementations. This userscript doesn't support older browsers, so it doesn't need to rely on jquery.